Docker: Add Dependabot, use Debian `stable` images, remove `${keyring_package}`
We weren't updating our Docker base images using Dependabot so do that.
Additionally, we can use `stable` by name so we roll over when new
stable releases are created, and there was a leftover
`${keyring_package}` substitution from an old version which can be
dropped.

The Ubuntu base image is left untouched deliberately to see if
Dependabot picks it up.